Create new go folder for go library "github.com/globalsign/mgo"
This new folder is a copy of the go folder for the "gopkg.in/mgo" library, with the code modified to work with a later and more up to date version of the mongodb go library "github.com/globalsign/mgo".
This folder also has a Docker file that builds an image for mongo 3.6.17.
See the README.md file for more details.