Get MapReduce partially working
Get simple MapReduce partially working, but its results still do not fully approximate to the results from the original javascript code.
The more Complex MapReduce example's results are not good at all.
Also keep the alternative code that produces the correct information for the search, to easily compare to the output from the erroneous MapReduce code.